Style : update directory
